+++ src/view.c
@@ -96,7 +96,7 @@ static void view_update_vy(View *v)
 {
     Window *w = v->window;
     int margin = window_get_scroll_margin(w);
-    int max_y = v->vy + w->edit_h - 1 - margin;
+    long max_y = v->vy + w->edit_h - 1 - margin;
 
     if (v->cy < v->vy + margin) {
         v->vy = v->cy - margin;
